implemented some mappings

+++ DrumMachine/RaspOutputController.h
@@ -32,7 +32,7 @@ public:
 
     void activeSampleChange(int newActiveSample) override;
 
-    void samplePlayPositionChange(vector<int>&) override;
+    void samplePlayPositionChange(vector<unsigned short>&) override;
 
     void blink(int);
